calc-angles-from-percent returns errors
for calc-angles-from-percent percents should always be less than 100 in
order for the last arc angle not to coinside or exceed the start angle
(for instanse, if one passes 25 and 75 percent and -Pi / 2 angle, the
angles will be -Pi / 2, 0, 3 * Pi / 2, -Pi / 2, that is the last two
angles have the same position on arc, which can lead to double line
drawn there). To catch this very possible situation, it now returns
error values too.

also refactored the tests for calc-angles-from-percent.